just rename the files totally
